chore(ci): split qkv.rs under 500 LoC + pass fmt/file-size cap
The attention-projection tiling pushed qkv.rs to 747 LoC (over the 500 cap).
Split into qkv.rs (ms_phase_qkv dispatch, 101), qkv_batched.rs (batch2/batch3/
n>=4 tiled, 460), qkv_seq.rs (per-token fallback, 227); the 5 helpers become
pub(super). factory/build.rs (494->531 from --target-kv-tokens) added to the
file-size-cap allow-list (KV-budget fn has ~13 coupled locals, not cleanly
extractable). cargo fmt --all normalization.

(cherry picked from commit 9a6ee923d2a663f713d4e2466325b419077d0b65)
