+++ crates/spark-model/src/layers/qwen3_attention/trait_impl/multi_seq/mod.rs
@@ -26,6 +26,8 @@ mod ffn;
 mod mla;
 mod mla_gemv;
 mod qkv;
+mod qkv_batched;
+mod qkv_seq;
 
 impl Qwen3AttentionLayer {
     #[allow(clippy::too_many_arguments)]
